layout adjustments

+++ app/src/main/java/br/com/dqrtech/currencylayerconverter/data/model/ConversionResponse.kt
@@ -1,9 +1,10 @@
 package br.com.dqrtech.currencylayerconverter.data.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 class ConversionResponse(
     success: Boolean,
     error: ErrorResponse?,
-    @field:SerializedName("quotes") val conversionRates: LinkedHashMap<String, Double>?
+    @field:SerializedName("quotes") val conversionRates: LinkedHashMap<String, BigDecimal>?
 ) : BaseResponse(success, error)
